requirements.txt
httpx>=0.27
python-dotenv>=1.0
pyyaml>=6
jsonschema>=4
jinja2>=3.1
psycopg[binary]>=3.2
